test: Implement Context
The test.Context implements ramen/e2e/test.Context. This is the way to
bind all the details needed for a single test. We implement only the
test.Context interface so we can use later e2e implementation, and keep
the code simple as possible. The ramenctl specific parts will be
implemented on top of the context.

Unlike ramen/e2e/test.Context, we keep a Command instead of the env and
config. This matches the actual model - all tests share the same config
and env, which are read-only at the point we use them.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>