Test clean command
This is the same as the run command, but we run only the unprotect and
undeploy steps, and clean the environment from test artifacts.

This does not remove all test namespaces yet. This should be fixed in
ramen/e2e.

The clean command logs to `report/test-clean.log`. If we run both
`test-run` and `test-clean` with the same output directory, we will get:

    % tree rdr
    rdr
    ├── test-clean.log
    └── test-run.log

Signed-off-by: Nir Soffer <nsoffer@redhat.com>