Change testing interface
Previously if a test step (setup, cleanup, test) failed, we returned an
error to signal the failure. This aborts the entire test and the
top level command code was logging the error to the console and exit.

This scheme does not work for complicated flow such as running multiple
test in parallel, of for handling multiple errors (e.g. failure to write
a test report after test step failed).

Change the interface so failed step log the error to the console and
logfile and return false, signaling that the operation failed and we
need to abort the test.

Previously we log the test success message in the top level command, but
this does not allow logging useful info when tests are finished, such as
number of passed and failed tests. Move the log to the test package.

When a test complete, it returns either a Command.Failed() error, or
call Command.Passed(). These methods will be used to write the test
report and report more info about the test.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>