Create test report for test run and clean commands
Since both run and clean commands need to store a report create a report
when creating a command.

Modify Setup(), Cleanup() and RunTest() to update the report when the
operation completes.

Modify Failed() and Passed() to write the report to the output
directory, and report test summary in the failure or success message.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>