test: Make report more generic
Use generic AddStep() instead of AddSetup() and AddCleanup(). This
allows adding more steps as needed, in the same way we use
console.Step().

A useful side effect of this change is validating that we don't add
duplicate step in AddStep(). Previously it was done in AddSetup() and
AddCleanup(), but if AddTest() was broken and added the "test" step
twice, this would not fail.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>