test: Simplify running test flows
Previously we added each test to the report in the goroutine running the
test, so we had to protect the report with a mutex. Simplify by keeping
test list in the command, and adding the tests to the report when all
tests completed.

Keeping test list enables adding new operations the test flow:

- Validating that tests can run on the cluster. For example disapp
  deployer using cephfs cannot run if not enabled in ramen config.
  Another example, vm workload cannot run if kubevirt is not installed.

- Gather data related to failed tests. We want to do this when all tests
  finished to avoid gathering load during the tests. This is also the
  simplest and most efficient way to gather data.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>