gather: Introduce the gather package
The gather package uses kubectl-gather[1] to gather namespaces from all
clusters. This will be used to gather data for failed tests, and later
for the `kubectl gather` command.

Since gathering data is best effort we don't return an error if
gathering fails. We may want to enhance this later for the gather
command.

Cluster data is stored in a cluster directory inside the output
directory:

    % tree -L2 test
    test
    ├── test-clean.log
    ├── test-clean.yaml
    ├── test-run.gather
    │   ├── dr1
    │   ├── dr2
    │   └── hub
    ├── test-run.log
    └── test-run.yaml

We gather entire namespaces, including some related cluster scope
resources. For more details check kubectl-gather[1].

[1] https://pkg.go.dev/github.com/nirs/kubectl-gather

Signed-off-by: Nir Soffer <nsoffer@redhat.com>