build: Upload built executables
Upload the built executables so they are available for download from the
build page. This will be useful for testing PRs by non-developers,
saving the time to build and send files manually.

We use a special tag "build-14503339779-1", corresponding to
https://github.com/RamenDR/ramenctl/actions/runs/14503339779/attempts/1.

The build tag is reported in --version:

    $ ramenctl --version
    build-14503339779-1

And in the report build info:

    build:
      version: build-14503339779-1
      commit: c4d07bacffdbd7c3ce4e13392f6521a613e80cf6

On macOS executables downloaded by the browser are considered unsafe and
cannot run. To remove the protection run:

    xattr -d com.apple.quarantine ramenctl

Fixes: #117
Signed-off-by: Nir Soffer <nsoffer@redhat.com>