test: Make clean flow more robust
When deleting the VRG ramen may wait until the VR reaches the desired
state. If storage is broken this may never happen, and deleting the VRG
may never complete. In the case describe in he issue, manual deletion of
the VR was required to allow test clean to complete.

Change the clean flow to first undeploy the application, which delete
the PVCs. This makes the VR less likely to get stuck because of
rbd-mirroring issues.

Undeploying the application deletes the placement and trigger a DRPC
deletion in ramen. This should be enough for cleaning up, however
undeploy does not wait until the DRPC is deleted in all cases, so we
call unprotect to ensure that the DRPC is deleted when the clean flow
completes.

Fixes: #112
Signed-off-by: Nir Soffer <nsoffer@redhat.com>