test: Use step helpers for tests step
The tests step cannot use failStep() or passStep() because it does not
run anything. It does not need to log anything to the console since
tests are reporting failures. The only thing we need is logging start
and finish events.

Add finishStep() helper for logging completion and adding the current
step to the report. With this the tests steps can use startStep() and
finishStep().

Signed-off-by: Nir Soffer <nsoffer@redhat.com>