test: Enable cancellation
Create a cancellable context when creating command.Command. The command
starts a signal handler that will cancel the context when receiving a
SIGINT signal. Cancelling the context will cancel the e2e calls and mark
the report as canceled.

Command.Stop() was introduced to stop the signal handler. This is useful
mostly for test that want to create a command.Command and clean up after
a test complete. It not required to stop the command on exit, but we do
this to serve as a good example.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>