test: Move run and clean flow to Command
The right way to test is the Command level, where we can configure the
test to work without real clusters. However the test flow is implemented
in test.Run() and test.Clean(), to we to duplicate this code in every
test.

Move the run flow from test.Run() to Command.Run(), and the clean flow
form test.Clean() to Command.Clean(). This makes test.Run() and
test.Clean() very simple, only wiring up things.

This will make the tests much simpler and increase code coverage.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>