test: Add GatherData option
We want to gather data by default, but during the tests we cannot gather
data since we don't have clusters to gather from. We probably want to be
able to disabling gathering when user prefer to gather data explicitly.

Make this possible by adding command Options with GatherData option. The
option is enabled by test.Run() and test.Clean(). The test do not need
to disable it since it not enabled by default.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>