test: Add deadlines for command operations
Add a 30 seconds deadline to validate and setup. These operations take
less than a second and we can safely fail quickly if they do not
complete in time.

Add a 1 minute deadline for cleanup. This takes usually 5 seconds. This
is the same timeout used in current ramen/2e. Next ramen/e2e will use
the deadline passed via the context.

This change is required to consume ramen/e2e context deadline added in
https://github.com/RamenDR/ramen/pull/2043

Signed-off-by: Nir Soffer <nsoffer@redhat.com>