test: Add Context.WithTimeout
Previously we returned parent.Context(), but we need to keep our own
context to be able to create a test.Context with a deadline.

Context.WithTimeout() returns a derive context with a deadline. The
context will fail with context.DeadlineExceeded on timeout.

This change is required to consume ramen/e2e context deadline added in
https://github.com/RamenDR/ramen/pull/2043

Signed-off-by: Nir Soffer <nsoffer@redhat.com>