test: Drop Step.Config
Test config is nice to have, but it makes it harder to create a generic
config. To keep it simple, drop the test config for now, so we can move
the Step type and step management code to the generic report.

This also makes the report little bit easier to scan since the config
blocks were very verbose. Locating the tests is still easy because the
test name is based on the config.

If we have a need later we can think how to provide this information in
a generic way.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>