validate: Implement config validation
Like test.Command, validate.Command is accepting a backend for
performing the actual work. This allows testing the command with a mock
backend without access to real clusters. Testing the actual validation
must be done manually and will be tested by end to end tests later.

Like test.Command we implement the Context interface to decouple the
command from the validation package.

Like test.Command we limit the time we wait for validating the config.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>