commands: Add drpc flags
Commands working with applications like `gather application` or
`validate application` need drpc name and namespace. Add the variables
in the root command, and a addDRPCFlags() helper to add the flags to the
commands.

Rename addOuputFlag() to addOutputFlags() for consistency. We may want
to add additional output flags in the future.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>