validation: Add Gather() interface
To write command tests we need to be able to mock the gather operation.
Add Gather() interface that can be mocked to access input arguments and
return mocked results.

This interface is relatively complicated, but we want to gather data
from all clusters in parallel and access the results as they become
available to update progress.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>