validate: Split application and clusters code
The validate commands share the validate.Command and code for validation
configuration, gathering namespaces and building the report. This code
remains in command.go.

We will need to add lot of specific code for validating gathered
clusters data or gathered application data. Move the clusters and
applications specific code to new file to make it easier to work with
the code.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>