validate: Show condition message on errors
Introduce report.Validate* types for validating conditions and other
values. We use report.ValidatedCondition to show validation result. The
other types are not use yet.

Since we need to show 2 or 3 properties for conditions, switch to
simpler []report.ValidatedCondition format.

If a condition is ok, we show:

    conditions:
    - type: Available
      state: ok ✅

If the condition is stale we show:

    conditions:
    - type: Available
      state: stale 🟠
      description: Observed generation 4 does not match object generation 5

If the condition has unexpected value we show the condition message.
Ramen conditions typically include useful message explaining why the
condition is not met:

    conditions:
    - type: Protected
      state: error ❌
      description: VolumeReplicationGroup (openshift-dr-ops/rhel-9-vm-drpc) on cluster
        prsurve-c1-7j is reporting errors (VolumeReplicationGroup is not in the admin
        namespace: VolumeReplicationGroup is not allowed to protect namespaces) readying
        workload data, retrying till DataReady condition is met

Since we use list of conditions we show the resources conditions in the
same order they appear in the ramen resources, instead of sorted by
type.

Fixes: #261
Signed-off-by: Nir Soffer <nsoffer@redhat.com>