validate: Validate drpc progression
Like phase, we mark only the stable progression as ok, and anything else
as an error. An application should now be in unstable progression for
long time. Possible case may be waiting for user action during failover
or relocate, or application stuck in failover or relocate because some
issue.

Example report:

    applicationStatus:
      hub:
        drpc:
          action:
            state: ok ✅
          conditions:
          - state: ok ✅
            type: Available
          - state: ok ✅
            type: PeerReady
          - state: ok ✅
            type: Protected
          deleted:
            state: ok ✅
          drPolicy: dr-policy
          name: appset-deploy-rbd
          namespace: argocd
          phase:
            state: ok ✅
            value: Deployed
          progression:
            state: ok ✅
            value: Completed

Part-of: #262
Signed-off-by: Nir Soffer <nsoffer@redhat.com>