helpers: Add UnifiedDiff function
The function accept expected and actual objects, and return a unified
diff between the objects YAML. If the objects are strings we compute a
diff of the strings since since marshaling yaml of multi-line string
return a different string.

This will be used by tests to show difference between actual and
expected objects.

Fixes #335

Signed-off-by: Nir Soffer <nsoffer@redhat.com>